Show which page the user is on
The app was set to only show the home page as the active page

Now javascript is run to automatically set the active page to whichever
page was loaded. This works but only if the user loads javascript.

It would be nice to do it using template language somehow in python but
I don't know how yet, so this will do.