Fix front end caching
It was caching in the incorrect order, and wasn't working on static site

I've fixed it so that it works, but it requires that the "cache"
attribute is set last, which is bad design. To fix the issue requires
reworking the component so that state is managed independantly of the
render, and the render method being called on each state change.

This wouldn't be too difficult, but it's a lot harder without
lit-element, and there are more important things to work on first, so
leaving this for now.